tango_with_django_project/requirements.txt
Cython==0.19.2
Django==1.7
Jinja2==2.7.1
MadMimi==1.1
MarkupSafe==0.18
Pillow==2.6.1
PyYAML==3.11
Pygments==1.6
South==1.0.1
Sphinx==1.1.3
argparse==1.2.1
django-appconf==0.6
django-autocomplete-light==2.0.5
django-guardian==1.2.5
django-haystack==2.3.1
django-image-cropping==0.9.1
django-registration-redux==1.1
django-selectable==0.9.0
django-widget-tweaks==1.3
docutils==0.11
easy-thumbnails==2.2
nose==1.3.0
numpy==1.8.0
pandas==0.12.0
pbr==0.10.0
pyenchant==1.6.5
pygame==1.9.2pre
pysolr==3.2.0
python-magic==0.4.6
requests==2.4.3
scipy==0.13.1
six==1.8.0
stevedore==1.1.0
sympy==0.7.3
virtualenv==1.11.6
virtualenv-clone==0.2.5
virtualenvwrapper==4.3.1
wsgiref==0.1.2
